pyproject.toml
[tool.poetry]
name = "hbv"
version = "0.1.0"
description = ""
authors = ["Kaiming <mads396@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
biopython = "^1.83"
scipy = "^1.11.4"
matplotlib = "^3.8.2"
numpy = "^1.26.3"
seaborn = "^0.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
